# Back-end/StudyApp.Api/Dtos/Auth/UserDto.cs
namespace StudyApp.Api.Dtos.Auth
{
    public class UserDto
    {
        public string? Id { get; set; }
        public string? UserName { get; set; }
        public string? Email { get; set; }
    }
}