# Back-end/StudyApp.Api/wwwroot/openapi.yaml
openapi: 3.0.3
info:
  title: StudyApp API
  version: 1.0.0
  description: OpenAPI spec for StudyApp (Auth, Flashcards, Knowledge Notes)
servers:
  - url: /
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [userName, email, password]
      properties:
        userName:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
    LoginRequest:
      type: object
      required: [userNameOrEmail, password]
      properties:
        userNameOrEmail:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        expiresIn:
          type: integer
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/UserDto"
    UserDto:
      type: object
      properties:
        id:
          type: string
        userName:
          type: string
        email:
          type: string
...
paths:
  /api/auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Created
        "400":
          description: Validation error
  /api/auth/login:
    post:
      summary: Login and receive JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Auth response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /api/auth/change-password:
    post:
      summary: Change current user's password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                currentPassword:
                  type: string
                newPassword:
                  type: string
      responses:
        "204":
          description: Password changed
...
security:
  - bearerAuth: []